Untested BLE code
